Refactor: config 폴더 삭제, constant 파일 정리
