+++ apps/api/src/auth/auth.constant.ts
@@ -3,3 +3,7 @@ export const jwtConstants = {
 };
 
 export const FORBIDDEN_USER_ROLE = '접근 권한 없음';
+
+// metadata keys
+export const IS_PUBLIC_KEY = 'isPublic';
+export const ROLE_KEY = 'roles';
