--- apps/api/src/auth/constants/metadata-keys.ts
@@ -1,2 +1,0 @@
-export const IS_PUBLIC_KEY = 'isPublic';
-export const ROLE_KEY = 'roles';
