+++ apps/api/src/image/image.constant.ts
@@ -1,5 +1,3 @@
-export const GITHUB_CLIENT_ID = 'GITHUB_CLIENT_ID';
-export const GITHUB_CLIENT_SECRET = 'GITHUB_CLIENT_SECRET';
 export const AWS_REGION = 'AWS_REGION';
 export const AWS_ACCESS_KEY = 'AWS_ACCESS_KEY';
 export const AWS_SECRET_KEY = 'AWS_SECRET_KEY';
