+++ apps/api/src/intra-auth/intra-auth.constant.ts
@@ -1,3 +1,6 @@
 export const SIGNIN_ALREADY_AUTH_ERROR_MESSAGE = '이미 인증된 사용자입니다.';
 export const CADET_ALREADY_EXIST_ERROR_MESSAGE = '이미 가입된 카뎃입니다.';
 export const NOT_EXIST_TOKEN_ERROR_MESSAGE = '존재하지 않는 토큰입니다.';
+
+export const TITLE = '[42WORLD] 이메일 인증을 완료하세요';
+export const EMAIL = 'student.42seoul.kr';
