+++ libs/common/src/filters/internal-server-error-exception.filter.ts
@@ -6,7 +6,7 @@ import {
 } from '@nestjs/common';
 import { Response } from 'express';
 import { errorHook } from '@api/utils';
-import { INTERNAL_ERROR_MESSAGE } from './constant';
+import { INTERNAL_ERROR_MESSAGE } from './filters.constant';
 
 @Catch(InternalServerErrorException)
 export class InternalServerErrorExceptionFilter implements ExceptionFilter {
