+++ libs/common/src/filters/typeorm-exception.filter.ts
@@ -3,7 +3,7 @@ import { Response } from 'express';
 import { TypeORMError } from 'typeorm';
 import { EntityNotFoundError } from 'typeorm/error/EntityNotFoundError';
 import { errorHook } from '@api/utils';
-import { INTERNAL_ERROR_MESSAGE } from './constant';
+import { INTERNAL_ERROR_MESSAGE } from './filters.constant';
 
 const FIND_DOUBLE_QUOTE = /\"/g;
